bagel/requirements.txt
datetime
azure-data-tables
azure-core
azure-storage-blob
pyyaml
pytest
pytest-cov
pandas
python-dotenv
datadog-api-client